pyproject.toml
[project]
name = "eq-distill"
version = "0.1.0"
description = "SAIR Mathematics Distillation Challenge — Equational Theories"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.40.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
